requirements.txt
# Core Framework
streamlit>=1.31.0

# Database
sqlalchemy>=2.0.0

# PDF Processing
pdfplumber>=0.10.0
pytesseract>=0.3.10
Pillow>=10.0.0

# Data Analysis
pandas>=2.0.0
numpy>=1.24.0

# Visualization
plotly>=5.18.0
altair>=5.2.0

# Utilities
python-dateutil>=2.8.2
pyyaml>=6.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
